requirements.txt
Django==5.2.14
djangorestframework==3.15.2
django-cors-headers==4.6.0
# psycopg v3 has pre-built wheels for newer Python (e.g. 3.14); psycopg2 often falls back to source builds.
psycopg[binary]==3.2.13
python-dotenv==1.1.1
django-extensions==4.1
pydot==3.0.4
pyparsing==3.2.3
reportlab==4.4.2
babel>=2.14
